move NotFoundTagType from model.dart to tag.dart
